Split scripts in App.vue to each js files
