+++ src/App.vue
@@ -22,224 +22,16 @@
 
 <script setup>
 // comp
-import { getRandom } from "./utils";
 import StatItem from "./components/StatItem.vue";
 import ToastList from "./components/ToastList.vue";
+// module
+import { refresh, cpu, ram, tx, rx, read, write } from "./puller";
+import { deleteToast, messages } from "./toast";
+import { SHOW_ALL_UI, SHOW_TOAST, CONTAINER_NAME } from "./user-props";
 
-import { ref, computed } from "vue";
 
-const FETCH_TIMEOUT = 20000;
-
-// user-properties variables (linked with WallpaperEngine)
-const ADDRESS = ref(process.env.VUE_APP_ADDRESS ?? "127.0.0.1"); // address
-const API_PORT = ref(process.env.VUE_APP_API_PORT ?? "1202"); // apiport
-const CONTAINER_NAME = ref(process.env.VUE_APP_CONTAINER_NAME ?? "docker-stats-api"); // containername
-const PULL_INTERVAL = ref(Number(process.env.VUE_APP_PULL_INTERVAL ?? 5000)); // pullinterval
-const SHOW_TOAST = ref(process.env.VUE_APP_SHOW_TOAST === "true"); // showtoast
-const STREAM_MODE = ref(process.env.VUE_APP_STREAM_MODE === "true"); // streammode
-const AGGREGATE = ref(process.env.VUE_APP_AGGREGATE === "true"); // aggregate
-const DEBUG_MODE = ref(process.env.VUE_APP_DEBUG_MODE === "true"); // debugmode
-
-// development options (only adjustable via .env file)
-const SHOW_ALL_UI = process.env.VUE_APP_SHOW_ALL_UI === "true";
-const RAND_STAT_VAL = process.env.VUE_APP_RAND_STAT_VAL === "true";
-
-// stats (ref)
-const cpu = ref(0);
-const ram = ref(0);
-const tx = ref(0);
-const rx = ref(0);
-const read = ref(0);
-const write = ref(0);
-
-// toast message system
-const toastIDStart = ref(0);
-const toastIDEnd = ref(0);
-const messages = ref({}); // used object to maintain order and state of toast components
-function makeToast(msg, lvl = 0) {
-    if (!SHOW_TOAST.value) return;
-    let icon = "";
-    switch (lvl) {
-        case 1:
-            icon = "💡 "; // info, tip
-            break;
-        case 2:
-            icon = "⚠️ "; // warn
-            break;
-        case 3:
-            icon = "🛑 "; // error
-            break;
-    }
-    messages.value[toastIDStart.value++] = icon + msg;
-}
-function deleteToast() {
-    delete messages.value[toastIDEnd.value++];
-    if (toastIDStart.value === toastIDEnd.value) {
-        toastIDStart.value = toastIDEnd.value = 0;
-    }
-}
-function clearAllToast() {
-    messages.value = {};
-    toastIDStart.value = toastIDEnd.value = 0;
-}
-
-// user-properties sync
-window.wallpaperPropertyListener = {
-    applyUserProperties: function (properties) {
-        if (properties.address) ADDRESS.value = properties.address.value;
-        if (properties.apiport) API_PORT.value = properties.apiport.value;
-        if (properties.containername) CONTAINER_NAME.value = properties.containername.value;
-        if (properties.pullinterval) {
-            let interval = Number(properties.pullinterval.value);
-            if (isNaN(interval)) makeToast(`${properties.pullinterval.value} is not a digit!`, 3);
-            else if (interval < 500) makeToast(`Pull interval cannot be lower than 500ms to prevent unexpected operation. The new value ${interval}ms will be ignored.`, 2);
-            else {
-                PULL_INTERVAL.value = interval;
-                if (interval < 3000) makeToast(`Frequent pulling(${PULL_INTERVAL.value}ms) can cause performance degradation of your server, increase network usage, and occur fetch timeoue error.`, 2);
-            }
-        }
-        if (properties.showtoast) {
-            SHOW_TOAST.value = properties.showtoast.value;
-            if (!SHOW_TOAST.value) clearAllToast();
-        }
-        if (properties.debugmode) {
-            DEBUG_MODE.value = properties.debugmode.value;
-            if (DEBUG_MODE.value) makeToast("Debug Mode consumes a lot of resources. This can cause fetch timeout error.", 2);
-        }
-        if (properties.streammode) {
-            STREAM_MODE.value = properties.streammode.value;
-            if (STREAM_MODE.value) makeToast("Stream mode is on 🚿", 1);
-            else makeToast("One-shot mode (legacy mode) is on. Turn off this only when your interval is bigger than 5 seconds.", 2);
-        }
-        if (properties.aggregate) AGGREGATE.value = properties.aggregate.value;
-        /* Add additional properties here when you customize this. */
-
-        if (DEBUG_MODE.value) makeToast(`Properties changed.\n${JSON.stringify(properties)}`, 1);
-        refreshPuller();
-    },
-};
-
-// fetch stats data
-let puller;
-let streamPuller = null;
-let prevTX = 0, prevRX = 0, prevRead = 0, prevWrite = 0;
-const API_URL = ref(`http://${ADDRESS.value}:${API_PORT.value}/${CONTAINER_NAME.value}`);
-if (RAND_STAT_VAL) {
-    puller = setInterval(() => {
-        cpu.value = getRandom(0, 100);
-        ram.value = getRandom(0, 1099511627776); // ~ 1TB
-        tx.value = getRandom(0, 1073741824); // ~ 1GB
-        read.value = getRandom(0, 1048576); // ~ 1MB
-    }, 1000);
-}
-else refreshPuller();
-
-function refreshPuller() {
-    API_URL.value = `http://${ADDRESS.value}:${API_PORT.value}/${CONTAINER_NAME.value}`;
-    clearInterval(puller);
-    streamPuller?.close();
-
-    if (STREAM_MODE.value) {
-        streamPuller = fetchStreamData();
-        puller = setInterval(() => {
-            if (streamPuller === null) streamPuller = fetchStreamData();
-        }, 5000);
-    }
-    else puller = setInterval(fetchData, PULL_INTERVAL.value);
-}
-function fetchData() {
-    if (DEBUG_MODE.value) makeToast(`GET > ${API_URL.value}`, 1);
-    fetch(API_URL.value, { signal: AbortSignal.timeout(FETCH_TIMEOUT) })
-        .then((res) => res.json())
-        .then(extractAndApplyValue)
-        .catch((err) => {
-            makeToast(`Error on fetch : ${err}`, 3);
-        });
-}
-function fetchStreamData() {
-    const STREAM_URL = `http://${ADDRESS.value}:${API_PORT.value}/stream/${CONTAINER_NAME.value}`;
-    const eventSource = new EventSource(STREAM_URL);
-    if (DEBUG_MODE.value) makeToast(`Open SSE connection > ${STREAM_URL}`, 1);
-
-    eventSource.onmessage = (event) => {
-        let stat;
-        try {
-            stat = JSON.parse(event.data);
-            extractAndApplyValue(stat);
-        }
-        catch (error) {
-            makeToast(error, 3);
-            return;
-        }
-    }
-
-    eventSource.onerror = (error) => {
-        makeToast(`Stream error occured.`, 3);
-        eventSource.close();
-        streamPuller = null;
-    }
-
-    return eventSource;
-}
-function extractAndApplyValue(stat) {
-    if (DEBUG_MODE.value) makeToast("< Fetched", 1);
-    
-    // Catch API error
-    if (stat.error != null) {
-        makeToast(stat, 3);
-        return;
-    }
-
-    // CPU
-    cpu.value = 0;
-    let cpuDelta = stat.cpu_stats.cpu_usage.total_usage - stat.precpu_stats.cpu_usage.total_usage;
-    let sysDelta = stat.cpu_stats.system_cpu_usage - stat.precpu_stats.system_cpu_usage;
-    let onlineCpu = stat.cpu_stats.online_cpus ?? 1;
-
-    if (cpuDelta > 0 && sysDelta > 0) {
-        cpu.value = (cpuDelta / sysDelta) * onlineCpu * 100;
-    }
-
-    // RAM
-    ram.value = stat.memory_stats.usage ?? 0;
-
-    // TX/RX
-    let txSum = 0,
-        rxSum = 0;
-    for (const adapter in stat.networks) {
-        txSum += stat.networks[adapter].tx_bytes ?? 0;
-        rxSum += stat.networks[adapter].rx_bytes ?? 0;
-    }
-    if (AGGREGATE.value) {
-        tx.value = txSum;
-        rx.value = rxSum;
-    } else {
-        tx.value = txSum - prevTX;
-        rx.value = rxSum - prevRX;
-        prevTX = txSum;
-        prevRX = rxSum;
-    }
-
-    // I/O
-    let ioStat = stat.blkio_stats.io_service_bytes_recursive;
-    if (ioStat === null) {
-        read.value = 0;
-        write.value = 0;
-        prevRead = 0;
-        prevWrite = 0;
-    }
-    else if (AGGREGATE.value) {
-        read.value = ioStat[0]?.value ?? 0;
-        write.value = ioStat[1]?.value ?? 0;
-    } else {
-        read.value = Math.max((ioStat[0]?.value ?? 0) - prevRead, 0);
-        write.value = Math.max((ioStat[1]?.value ?? 0) - prevWrite, 0);
-        prevRead = ioStat[0]?.value ?? 0;
-        prevWrite = ioStat[1]?.value ?? 0;
-    }
-
-    if (DEBUG_MODE.value) makeToast("Value updated", 1);
-}
+// start puller
+refresh();
 </script>
 
 <style>
